requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
bottleneck=1.3.7=py39h9128911_0
brotli-python=1.0.9=py39hd77b12b_7
ca-certificates=2024.2.2=h56e8100_0
cachetools=5.3.2=pypi_0
captum=0.7.0=pypi_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=py39h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.1=pyhd8ed1ab_0
configparser=6.0.0=pypi_0
contourpy=1.2.0=pypi_0
cryptography=41.0.7=py39h89fc84f_0
cuda-cccl=12.3.101=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.101=0
cuda-opencl-dev=12.3.101=0
cuda-profiler-api=12.3.101=0
cuda-runtime=12.1.0=0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py39hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
dm-tree=0.1.8=pypi_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=py39haa95532_0
flatbuffers=23.5.26=pypi_0
fonttools=4.47.2=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2024.2.0=pypi_0
gast=0.5.4=pypi_0
giflib=5.2.1=h8cc25b3_3
gmpy2=2.1.2=py39h7f96b67_0
google-auth=2.27.0=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.60.0=pypi_0
h5py=3.10.0=pypi_0
idna=3.4=py39haa95532_0
imageio=2.34.0=pypi_0
imbalanced-learn=0.12.0=pypi_0
imblearn=0.0=pypi_0
importlib-metadata=7.0.1=pyha770c72_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=7.0.1=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.0=pyha63f2e9_0
ipython=8.12.0=pyh08f2357_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=py39haa95532_0
joblib=1.3.2=pypi_0
jpeg=9e=h2bbff1b_1
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=4.12.0=py39hcbf5309_0
keras=2.15.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
kymatio=0.3.0=pypi_0
lazy-loader=0.3=pypi_0
lerc=3.0=hd77b12b_0
libclang=16.0.6=pypi_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.4.107=0
libcurand-dev=10.3.4.107=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lime=0.2.0.1=pypi_0
llvmlite=0.42.0=pypi_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py39h2bbff1b_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
ml-dtypes=0.2.0=pypi_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py39haa95532_0
namex=0.0.7=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=py39haa95532_0
numba=0.59.0=pypi_0
numexpr=2.8.7=py39h2cd9be0_0
numpy=1.26.3=py39h055cbcc_0
numpy-base=1.26.3=py39h65a83cf_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.13=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.0.1=py39h045eedc_0
pip=23.3.1=py39haa95532_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=4.23.4=pypi_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pyhd8ed1ab_0
pyopenssl=23.2.0=py39haa95532_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=2.1.2=py3.9_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyvips=2.2.2=pypi_0
pywin32=227=py39hb82d6ee_1
pyyaml=6.0.1=py39h2bbff1b_0
pyzmq=25.1.2=py39hd77b12b_0
requests=2.31.0=py39haa95532_0
requests-oauthlib=1.3.1=pypi_0
rich=13.7.0=pypi_0
rsa=4.9=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.1=pypi_0
setuptools=68.2.2=py39haa95532_0
shap=0.44.1=pypi_0
six=1.16.0=pyh6c4a22f_0
slicer=0.0.7=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py39haa95532_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.15.0=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-intel=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2024.1.30=pypi_0
tk=8.6.12=h2bbff1b_0
torchaudio=2.1.2=pypi_0
torchinfo=1.8.0=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.2=py39hb82d6ee_0
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
typing_extensions=4.9.0=py39haa95532_1
tzdata=2023.4=pypi_0
urllib3=1.26.18=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wrapt=1.14.1=pypi_0
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
zeromq=4.3.5=hd77b12b_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
